io.Source is not closed when used in most common text reading idiom - Provide new tools for reading in text   - from a file   - from a string file name   - from a resource file - text can be read in as   - a single string with newlines   - a seq of strings, one string per line - FileUtils put in its own file   - in same package to keep existing API the same
Hopefully this will protect users from creating resource leaks that
may explain sbt crashing.
Also simpler interface should make code more streamlined, existing
uses of io.Source seem often to be converting between text and seq
unnecessarily.
